Added sound effects for firing bullets and blast

+++ src/Alien.cpp
@@ -5,15 +5,15 @@ Alien::Alien(IMeshSceneNode *node, AlienType type){
 }
 
 void Alien::moveLeft(float speed){
-	alienNode->setPosition(alienNode->getPosition()+speed*vector3df(-0.4,0,0));
+	alienNode->setPosition(alienNode->getPosition()+speed*vector3df((f32)-0.4,0,0));
 }
 
 void Alien::moveRight(float speed){
-	alienNode->setPosition(alienNode->getPosition()+speed*vector3df(0.4,0,0));
+	alienNode->setPosition(alienNode->getPosition()+speed*vector3df((f32)0.4,0,0));
 }
 
 void Alien::moveDown(){
-	alienNode->setPosition(alienNode->getPosition()+vector3df(0,-5.0,0));
+	alienNode->setPosition(alienNode->getPosition()+vector3df(0,(f32)-5.0,0));
 }
 
 void Alien::destroy(){
